feat(wasm-utxo): isolate WASM bindings into separate module
Move all WASM-specific code into a dedicated module to improve
organization and separation of concerns. The core functionality remains
unchanged, but the binding layer is now cleanly separated from the
implementation.

This refactoring:
- Creates a new `wasm` module to contain all bindings
- Moves JS/WASM interface code out of core implementation files
- Maintains backward compatibility with existing JS API

Issue: BTC-2652
