+++ packages/wasm-utxo/src/lib.rs
@@ -1,12 +1,8 @@
 mod address;
-mod descriptor;
 mod error;
-mod fixed_script_wallet;
-mod miniscript;
+pub mod fixed_script_wallet;
 mod networks;
-mod psbt;
-mod try_from_js_value;
-mod try_into_js_value;
+pub mod wasm;
 
 // re-export bitcoin from the miniscript crate
 // this package is transitioning to a all-purpose bitcoin package, so we want easy access
@@ -17,9 +13,7 @@ pub use address::{
     to_output_script_with_network, utxolib_compat,
 };
 
-pub use descriptor::WrapDescriptor;
-pub use miniscript::WrapMiniscript;
 pub use networks::Network;
-pub use psbt::WrapPsbt;
+pub use wasm::{WrapDescriptor, WrapMiniscript, WrapPsbt};
 
 pub use crate::fixed_script_wallet::*;
