+++ packages/wasm-utxo/src/wasm/fixed_script_wallet.rs
@@ -1,0 +1,66 @@
+use wasm_bindgen::prelude::*;
+use wasm_bindgen::JsValue;
+
+use crate::address::networks::AddressFormat;
+use crate::address::utxolib_compat::UtxolibNetwork;
+use crate::error::WasmUtxoError;
+use crate::fixed_script_wallet::{Chain, WalletScripts};
+use crate::wasm::try_from_js_value::TryFromJsValue;
+use crate::wasm::wallet_keys_helpers::root_wallet_keys_from_jsvalue;
+
+#[wasm_bindgen]
+pub struct FixedScriptWalletNamespace;
+
+#[wasm_bindgen]
+impl FixedScriptWalletNamespace {
+    #[wasm_bindgen]
+    pub fn output_script(
+        keys: JsValue,
+        chain: u32,
+        index: u32,
+        network: JsValue,
+    ) -> Result<Vec<u8>, WasmUtxoError> {
+        let network = UtxolibNetwork::try_from_js_value(&network)?;
+        let chain = Chain::try_from(chain)
+            .map_err(|e| WasmUtxoError::new(&format!("Invalid chain: {}", e)))?;
+
+        let wallet_keys = root_wallet_keys_from_jsvalue(&keys)?;
+        let scripts = WalletScripts::from_wallet_keys(
+            &wallet_keys,
+            chain,
+            index,
+            &network.output_script_support(),
+        )?;
+        Ok(scripts.output_script().to_bytes())
+    }
+
+    #[wasm_bindgen]
+    pub fn address(
+        keys: JsValue,
+        chain: u32,
+        index: u32,
+        network: JsValue,
+        address_format: Option<String>,
+    ) -> Result<String, WasmUtxoError> {
+        let network = UtxolibNetwork::try_from_js_value(&network)?;
+        let wallet_keys = root_wallet_keys_from_jsvalue(&keys)?;
+        let chain = Chain::try_from(chain)
+            .map_err(|e| WasmUtxoError::new(&format!("Invalid chain: {}", e)))?;
+        let scripts = WalletScripts::from_wallet_keys(
+            &wallet_keys,
+            chain,
+            index,
+            &network.output_script_support(),
+        )?;
+        let script = scripts.output_script();
+        let address_format = AddressFormat::from_optional_str(address_format.as_deref())
+            .map_err(|e| WasmUtxoError::new(&format!("Invalid address format: {}", e)))?;
+        let address = crate::address::utxolib_compat::from_output_script_with_network(
+            &script,
+            &network,
+            address_format,
+        )
+        .map_err(|e| WasmUtxoError::new(&format!("Failed to generate address: {}", e)))?;
+        Ok(address.to_string())
+    }
+}
