+++ packages/wasm-utxo/src/wasm/mod.rs
@@ -1,0 +1,17 @@
+mod address;
+mod bip32interface;
+mod descriptor;
+mod fixed_script_wallet;
+mod miniscript;
+mod psbt;
+mod try_from_js_value;
+mod try_into_js_value;
+mod utxolib_compat;
+pub(crate) mod wallet_keys_helpers;
+
+pub use address::AddressNamespace;
+pub use descriptor::WrapDescriptor;
+pub use fixed_script_wallet::FixedScriptWalletNamespace;
+pub use miniscript::WrapMiniscript;
+pub use psbt::WrapPsbt;
+pub use utxolib_compat::UtxolibCompatNamespace;
